requirements.txt
pillow==10.3.0
argparse==1.4.0
